+++ linux/library/src/method_result.cc
@@ -11,37 +11,11 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-#include "linux/library/include/flutter_desktop_embedding/channels.h"
+#include "linux/library/include/flutter_desktop_embedding/method_result.h"
 
 #include <iostream>
 
 namespace flutter_desktop_embedding {
-
-constexpr char kMessageMethodKey[] = "method";
-constexpr char kMessageArgumentsKey[] = "args";
-
-MethodCall::MethodCall(const std::string &method_name,
-                       const Json::Value &arguments)
-    : method_name_(method_name), arguments_(arguments) {}
-
-MethodCall::~MethodCall() {}
-
-std::unique_ptr<MethodCall> MethodCall::CreateFromMessage(
-    const Json::Value &message) {
-  Json::Value method = message[kMessageMethodKey];
-  if (method.isNull()) {
-    return nullptr;
-  }
-  Json::Value arguments = message[kMessageArgumentsKey];
-  return std::make_unique<MethodCall>(method.asString(), arguments);
-}
-
-Json::Value MethodCall::AsMessage() const {
-  Json::Value message(Json::objectValue);
-  message[kMessageMethodKey] = method_name_;
-  message[kMessageArgumentsKey] = arguments_;
-  return message;
-}
 
 void MethodResult::Success(const Json::Value &result) {
   SuccessInternal(result);
